% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oecd.R
\name{oecd}
\alias{oecd}
\title{Download OECD data}
\usage{
oecd(country = "all", indicator = "QASA_TABLE7PSD", start = "2000-Q1",
  end = "2011-Q4", extra = FALSE)
}
\arguments{
\item{country}{character vector of country names in ISO 3 country code
format. The default (\code{'all'}) downloads data for all available
countries.}

\item{indicator}{character vector of indicator codes.}

\item{start}{first year or quarter to download data for.}

\item{end}{final year or quarter to download data for.}

\item{extra}{logical. If \code{TRUE} then all information from the download
is returned.}
}
\value{
a data frame
}
\description{
Downloads requested data from the OECD's API, parses the download, and
returns the data in country-time format
}
\examples{
\dontrun{
qrt_public_debt <- oecd()
}
}

