% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{oecdr_connect}
\alias{oecdr_connect}
\title{Get a DBI connection from env (Postgres) or fallback to SQLite}
\usage{
oecdr_connect(drv = c("RSQLite", "RPostgres"))
}
\arguments{
\item{drv}{"RPostgres" or "RSQLite"}
}
\description{
Get a DBI connection from env (Postgres) or fallback to SQLite
}
