% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{write_db}
\alias{write_db}
\title{Write a data.frame to a DB table}
\usage{
write_db(df, table, con = NULL, overwrite = FALSE, append = TRUE)
}
\arguments{
\item{df}{data.frame/tibble}

\item{table}{table name}

\item{con}{optional DBI connection. If NULL, creates a SQLite temp DB.}

\item{overwrite}{overwrite table if exists}

\item{append}{append rows when table exists and overwrite=FALSE}
}
\description{
Write a data.frame to a DB table
}
