% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oecd.R
\name{oecd}
\alias{oecd}
\title{Download OECD data}
\usage{
oecd(country = "all", table = "QASA_TABLE7PSD",
  indicator = "SAFGD.S1311C.CAR.NSA", start = "2000-Q1", end = "2011-Q4",
  extra = FALSE, return_url = FALSE)
}
\arguments{
\item{country}{character vector of country names in ISO 3 country code
format. The default (\code{'all'}) downloads data for all available
countries.}

\item{table}{character string identifying the table. Note only one table
can be queried at a time.}

\item{indicator}{character vector of indicator codes for variables in
\code{table}.}

\item{start}{first year or quarter to download data for.}

\item{end}{final year or quarter to download data for.}

\item{extra}{logical. If \code{TRUE} then all information from the download
is returned.}

\item{return_url}{logical. Whether or not to return the URL constructed
for the query. Can be useful for debugging your query.}
}
\value{
a data frame with (if \code{extra = TRUE}) country ISO 3 Letter
codes, the time variable and the requested indicators.
}
\description{
Downloads requested data from the OECD's API, parses the download, and
returns the data in country-time format
}
\examples{
\dontrun{
# Download: Central Governement Currency and Deposits (\% GDP) and
#           Total Gross Central Government Debt in current prices
qrt_public_debt <- oecd(indicator = c('SAF2LXT.S1311C.PCTGDPA.NSA',
                     'SAFGD.S1311C.CAR.NSA'))

# Gross domestic product (expenditure approach current prices, millions)
gdp <- oecd(table = 'SNA_TABLE1', indicator = 'B1_GE.CPC',
            start = 2000, end = 2012)
}
}

